Cast invalid values.
git-svn-id: https://svn.dealii.org/trunk@6561 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/deal.II/source/fe/fe.cc
@@ -302,7 +302,8 @@ FiniteElementBase<dim>::interface_constraints_size () const
       default:
             Assert (false, ExcNotImplemented());
     };
-  return TableIndices<2> (-1, -1);
+  return TableIndices<2> (static_cast<unsigned int>(-1),
+                          static_cast<unsigned int>(-1));
 };
